Update door lock properties, align naming to ISY conventions
