+++ isy994/items/devices/common/device_lock.py
@@ -8,6 +8,8 @@ class Device_Lock(Device_Base):
     def __init__(self, container, name, address):
         Device_Base.__init__(self, container, "lock", name, address)
 
+        self.add_property("alarm", None)
+        self.add_property("alarmcode", 0)
         self.add_property("batterylevel", 0)
         self.add_property("state", 0)
         self.add_property("error", 0)
